requirements.txt
Flask
twilio
requests
openai
python-dotenv
